requirements.txt
buttplug-py>=0.1.0
tomli>=2.0.1
pydantic>=1.10.0,<2.0.0
quart>=0.18.0
quart-cors>=0.5.0 